requirements.txt
numpy==1.23.5
opencv-python==4.7.0.72
matplotlib==3.7.1
scipy==1.10.1
Pillow==9.5.0
imageio[ffmpeg]==2.22.4
scikit-learn==1.2.2
diffusers==0.16.1
huggingface_hub==0.20.2
transformers==4.30.2
